+++ config/routes.rb
@@ -1,6 +1,9 @@
 Rails.application.routes.draw do
   resources :projects
   resources :tasks
+  resources :user_sessions
   root to: "tasks#main"
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+
+  get 'login' => 'user_sessions#new', :as => :login
+  post 'logout' => 'user_sessions#destroy', :as => :logout
 end