+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# (not yet)
+
+- Require login
+
 # v0.1.0 (2018-08-15)
 
 - Initial release